Run functional tests without the doctrine-orm-infrastructure dependency
