+++ tests/Doctrine/TranslatableClassMetadataTest.php
@@ -8,9 +8,10 @@ use PHPUnit\Framework\TestCase;
 use Webfactory\Bundle\PolyglotBundle\Doctrine\TranslatableClassMetadata;
 use Webfactory\Bundle\PolyglotBundle\Tests\Fixtures\Entity\TestEntity;
 use Webfactory\Bundle\PolyglotBundle\Tests\Fixtures\Entity\TestEntityTranslation;
+use Webfactory\Bundle\PolyglotBundle\Tests\Functional\DatabaseFunctionalTestCase;
 use Webfactory\Doctrine\ORMTestInfrastructure\ORMInfrastructure;
 
-class TranslatableClassMetadataTest extends TestCase
+class TranslatableClassMetadataTest extends DatabaseFunctionalTestCase
 {
     /**
      * @test
@@ -27,11 +28,7 @@ class TranslatableClassMetadataTest extends TestCase
 
     private function createMetadata(): TranslatableClassMetadata
     {
-        $infrastructure = new ORMInfrastructure([
-            TestEntity::class,
-            TestEntityTranslation::class,
-        ], mappingDriver: new AttributeDriver([], true));
-        $entityManager = $infrastructure->getEntityManager();
+        $entityManager = $this->entityManager;
 
         return TranslatableClassMetadata::parseFromClass(TestEntity::class, $entityManager->getMetadataFactory());
     }
