+++ tests/Functional/CascadePersistTranslationsTest.php
@@ -2,38 +2,31 @@
 
 namespace Webfactory\Bundle\PolyglotBundle\Tests\Functional;
 
-use Doctrine\Common\Collections\ArrayCollection;
-use Doctrine\Common\Collections\Collection;
-use Doctrine\ORM\Mapping as ORM;
-use Webfactory\Bundle\PolyglotBundle\Attribute as Polyglot;
-use Webfactory\Bundle\PolyglotBundle\Translatable;
-use Webfactory\Bundle\PolyglotBundle\TranslatableInterface;
+use PHPUnit\Framework\Attributes\Test;
+use Webfactory\Bundle\PolyglotBundle\Tests\Fixtures\Entity\CascadePersist\CascadePersistTranslationsTest_Entity;
+use Webfactory\Bundle\PolyglotBundle\Tests\Fixtures\Entity\CascadePersist\CascadePersistTranslationsTest_Translation;
 
 /**
  * This test shows that `PersistentTranslatable` will take care of adding new
  * translation entities to the entity manager, so explicit `persist()` calls are not
  * necessary; and also, `cascade={"persist"}` isn't either.
  */
-class CascadePersistTranslationsTest extends FunctionalTestBase
+class CascadePersistTranslationsTest extends DatabaseFunctionalTestCase
 {
-    protected function setUp(): void
+    #[Test]
+    public function adding_and_persisting_translations(): void
     {
-        parent::setUp();
-        $this->setupOrmInfrastructure([
+        self::setupSchema([
             CascadePersistTranslationsTest_Entity::class,
             CascadePersistTranslationsTest_Translation::class,
         ]);
-    }
 
-    /**
-     * @test
-     */
-    public function adding_and_persisting_translations(): void
-    {
         $entity = new CascadePersistTranslationsTest_Entity();
         $entity->addTranslation('de_DE', 'text de_DE');
-        $this->entityManager->persist($entity);
-        $this->entityManager->flush();
+
+        self::import([$entity]);
+//        $this->entityManager->persist($entity);
+//        $this->entityManager->flush();
 
         // note the absent call to $this->entityManager->persist() for the translation entity
 
@@ -42,55 +35,4 @@ class CascadePersistTranslationsTest extends FunctionalTestBase
         self::assertCount(1, $result);
         self::assertSame('text de_DE', $result[0]['text']);
     }
-}
-
-#[Polyglot\Locale(primary: 'en_GB')]
-#[ORM\Entity]
-class CascadePersistTranslationsTest_Entity
-{
-    #[ORM\Column(type: 'integer')]
-    #[ORM\Id]
-    #[ORM\GeneratedValue]
-    private ?int $id = null;
-
-    /**
-     * (!) There is *not* cascade="persist" configuration here.
-     */
-    #[Polyglot\TranslationCollection]
-    #[ORM\OneToMany(targetEntity: \CascadePersistTranslationsTest_Translation::class, mappedBy: 'entity')]
-    protected Collection $translations;
-
-    #[Polyglot\Translatable]
-    #[ORM\Column(type: 'string')]
-    protected string|TranslatableInterface $text;
-
-    public function __construct()
-    {
-        $this->text = new Translatable('test en_GB');
-        $this->translations = new ArrayCollection();
-    }
-
-    public function addTranslation(string $locale, mixed $text): void
-    {
-        $this->text->setTranslation($text, $locale);
-    }
-}
-
-#[ORM\Entity]
-class CascadePersistTranslationsTest_Translation
-{
-    #[ORM\Id]
-    #[ORM\GeneratedValue]
-    #[ORM\Column(type: 'integer')]
-    private ?int $id = null;
-
-    #[Polyglot\Locale]
-    #[ORM\Column]
-    private string $locale;
-
-    #[ORM\ManyToOne(targetEntity: \CascadePersistTranslationsTest_Entity::class, inversedBy: 'translations')]
-    private CascadePersistTranslationsTest_Entity $entity;
-
-    #[ORM\Column]
-    private string $text;
 }
