+++ tests/Functional/EntityInheritanceTest.php
@@ -2,23 +2,23 @@
 
 namespace Webfactory\Bundle\PolyglotBundle\Tests\Functional;
 
-use Doctrine\Common\Collections\ArrayCollection;
-use Doctrine\Common\Collections\Collection;
-use Doctrine\ORM\Mapping as ORM;
-use Webfactory\Bundle\PolyglotBundle\Attribute as Polyglot;
+use Webfactory\Bundle\PolyglotBundle\Tests\Fixtures\Entity\EntityInheritance\EntityInheritance_BaseEntityClass;
+use Webfactory\Bundle\PolyglotBundle\Tests\Fixtures\Entity\EntityInheritance\EntityInheritance_BaseEntityClassTranslation;
+use Webfactory\Bundle\PolyglotBundle\Tests\Fixtures\Entity\EntityInheritance\EntityInheritance_ChildEntityClass;
+use Webfactory\Bundle\PolyglotBundle\Tests\Fixtures\Entity\EntityInheritance\EntityInheritance_ChildEntityClassTranslation;
 use Webfactory\Bundle\PolyglotBundle\Translatable;
-use Webfactory\Bundle\PolyglotBundle\TranslatableInterface;
 
 /**
  * This tests translations for different fields in an inheritance hierarchy. For every
  * entity class in the hierarchy, a dedicated translations class has to be used.
  */
-class EntityInheritanceTest extends FunctionalTestBase
+class EntityInheritanceTest extends DatabaseFunctionalTestCase
 {
     protected function setUp(): void
     {
         parent::setUp();
-        $this->setupOrmInfrastructure([
+
+        self::setupSchema([
             EntityInheritance_BaseEntityClass::class,
             EntityInheritance_BaseEntityClassTranslation::class,
             EntityInheritance_ChildEntityClass::class,
@@ -37,9 +37,9 @@ class EntityInheritanceTest extends FunctionalTestBase
         $t2->setTranslation('Extratext', 'de_DE');
         $entity->setExtra($t2);
 
-        $this->infrastructure->import($entity);
+        self::import([$entity]);
 
-        $loaded = $this->infrastructure->getEntityManager()->find(EntityInheritance_ChildEntityClass::class, $entity->getId());
+        $loaded = $this->entityManager->find(EntityInheritance_ChildEntityClass::class, $entity->getId());
 
         self::assertSame('Basistext', $loaded->getText()->translate('de_DE'));
         self::assertSame('Extratext', $loaded->getExtraText()->translate('de_DE'));
@@ -49,11 +49,12 @@ class EntityInheritanceTest extends FunctionalTestBase
 
     public function testAddTranslation(): void
     {
-        $entityManager = $this->infrastructure->getEntityManager();
+        $entityManager = $this->entityManager;
+
         $entity = new EntityInheritance_ChildEntityClass();
         $entity->setText(new Translatable('base text'));
         $entity->setExtra(new Translatable('extra text'));
-        $this->infrastructure->import($entity);
+        self::import([$entity]);
 
         $loaded = $entityManager->find(EntityInheritance_ChildEntityClass::class, $entity->getId());
         $loaded->getText()->setTranslation('Basistext', 'de_DE');
@@ -61,6 +62,7 @@ class EntityInheritanceTest extends FunctionalTestBase
         $entityManager->flush();
 
         $entityManager->clear();
+
         $reloaded = $entityManager->find(EntityInheritance_ChildEntityClass::class, $entity->getId());
 
         self::assertSame('Basistext', $reloaded->getText()->translate('de_DE'));
@@ -71,7 +73,7 @@ class EntityInheritanceTest extends FunctionalTestBase
 
     public function testUpdateTranslations(): void
     {
-        $entityManager = $this->infrastructure->getEntityManager();
+        $entityManager = $this->entityManager;
 
         $entity = new EntityInheritance_ChildEntityClass();
         $t1 = new Translatable('old base text');
@@ -82,7 +84,7 @@ class EntityInheritanceTest extends FunctionalTestBase
         $t2->setTranslation('alter Extratext', 'de_DE');
         $entity->setExtra($t2);
 
-        $this->infrastructure->import($entity);
+        self::import([$entity]);
 
         $loaded = $entityManager->find(EntityInheritance_ChildEntityClass::class, $entity->getId());
         $loaded->getText()->setTranslation('new base text');
@@ -99,113 +101,4 @@ class EntityInheritanceTest extends FunctionalTestBase
         self::assertSame('new base text', $reloaded->getText()->translate('en_GB'));
         self::assertSame('new extra text', $reloaded->getExtraText()->translate('en_GB'));
     }
-}
-
-#[Polyglot\Locale(primary: 'en_GB')]
-#[ORM\Entity]
-#[ORM\InheritanceType(value: 'SINGLE_TABLE')]
-#[ORM\DiscriminatorMap(['base' => 'EntityInheritance_BaseEntityClass', 'child' => 'EntityInheritance_ChildEntityClass'])]
-#[ORM\DiscriminatorColumn(name: 'discriminator', type: 'string')]
-class EntityInheritance_BaseEntityClass
-{
-    #[ORM\Column(type: 'integer')]
-    #[ORM\Id]
-    #[ORM\GeneratedValue]
-    private ?int $id = null;
-
-    private string $discriminator;
-
-    #[Polyglot\TranslationCollection]
-    #[ORM\OneToMany(targetEntity: \EntityInheritance_BaseEntityClassTranslation::class, mappedBy: 'entity')]
-    private Collection $translations;
-
-    #[Polyglot\Translatable]
-    #[ORM\Column(type: 'string')]
-    private TranslatableInterface|string|null $text = null;
-
-    public function __construct()
-    {
-        $this->translations = new ArrayCollection();
-    }
-
-    public function getId(): int
-    {
-        return $this->id;
-    }
-
-    public function setText(TranslatableInterface $text): void
-    {
-        $this->text = $text;
-    }
-
-    public function getText(): TranslatableInterface
-    {
-        return $this->text;
-    }
-}
-
-#[ORM\Entity]
-class EntityInheritance_BaseEntityClassTranslation
-{
-    #[ORM\Id]
-    #[ORM\GeneratedValue]
-    #[ORM\Column(type: 'integer')]
-    private ?int $id = null;
-
-    #[Polyglot\Locale]
-    #[ORM\Column]
-    private string $locale;
-
-    #[ORM\ManyToOne(targetEntity: \EntityInheritance_BaseEntityClass::class, inversedBy: 'translations')]
-    private EntityInheritance_BaseEntityClass $entity;
-
-    #[ORM\Column]
-    private string $text;
-}
-
-#[ORM\Entity]
-class EntityInheritance_ChildEntityClass extends EntityInheritance_BaseEntityClass
-{
-    #[Polyglot\Translatable]
-    #[ORM\Column(type: 'string')]
-    private TranslatableInterface|string|null $extraText = null;
-
-    #[Polyglot\TranslationCollection]
-    #[ORM\OneToMany(targetEntity: \EntityInheritance_ChildEntityClassTranslation::class, mappedBy: 'entity')]
-    private Collection $extraTranslations;
-
-    public function __construct()
-    {
-        parent::__construct();
-        $this->extraTranslations = new ArrayCollection();
-    }
-
-    public function setExtra(TranslatableInterface $extraText): void
-    {
-        $this->extraText = $extraText;
-    }
-
-    public function getExtraText(): TranslatableInterface
-    {
-        return $this->extraText;
-    }
-}
-
-#[ORM\Entity]
-class EntityInheritance_ChildEntityClassTranslation
-{
-    #[ORM\Id]
-    #[ORM\GeneratedValue]
-    #[ORM\Column(type: 'integer')]
-    private ?int $id = null;
-
-    #[Polyglot\Locale]
-    #[ORM\Column]
-    private string $locale;
-
-    #[ORM\ManyToOne(targetEntity: \EntityInheritance_ChildEntityClass::class, inversedBy: 'extraTranslations')]
-    private EntityInheritance_ChildEntityClass $entity;
-
-    #[ORM\Column]
-    private string $extraText;
 }
