+++ tests/Functional/TranslatingUnmappedPropertiesTest.php
@@ -2,10 +2,8 @@
 
 namespace Webfactory\Bundle\PolyglotBundle\Tests\Functional;
 
-use Doctrine\Common\Collections\ArrayCollection;
-use Doctrine\Common\Collections\Collection;
-use Doctrine\ORM\Mapping as ORM;
-use Webfactory\Bundle\PolyglotBundle\Attribute as Polyglot;
+use Webfactory\Bundle\PolyglotBundle\Tests\Fixtures\Entity\UnmappedProperties\TranslatingUnmappedPropertiesTest_Entity;
+use Webfactory\Bundle\PolyglotBundle\Tests\Fixtures\Entity\UnmappedProperties\TranslatingUnmappedPropertiesTest_Translation;
 use Webfactory\Bundle\PolyglotBundle\Translatable;
 
 /**
@@ -15,12 +13,13 @@ use Webfactory\Bundle\PolyglotBundle\Translatable;
  * This is useful when these fields are managed or updated by e.g. lifecycle callbacks
  * or other Doctrine event listeners.
  */
-class TranslatingUnmappedPropertiesTest extends FunctionalTestBase
+class TranslatingUnmappedPropertiesTest extends DatabaseFunctionalTestCase
 {
     protected function setUp(): void
     {
         parent::setUp();
-        $this->setupOrmInfrastructure([
+
+        self::setupSchema([
             TranslatingUnmappedPropertiesTest_Entity::class,
             TranslatingUnmappedPropertiesTest_Translation::class,
         ]);
@@ -32,86 +31,11 @@ class TranslatingUnmappedPropertiesTest extends FunctionalTestBase
         $entity->text = new Translatable('base text');
         $entity->text->setTranslation('Basistext', 'de_DE');
 
-        $this->infrastructure->import($entity);
+        self::import($entity);
 
-        $loaded = $this->infrastructure->getEntityManager()->find(TranslatingUnmappedPropertiesTest_Entity::class, $entity->id);
+        $loaded = $this->entityManager->find(TranslatingUnmappedPropertiesTest_Entity::class, $entity->id);
 
         self::assertSame('Basistext', $loaded->text->translate('de_DE'));
         self::assertSame('base text', $loaded->text->translate('en_GB'));
-    }
-}
-
-#[ORM\Entity]
-#[ORM\HasLifecycleCallbacks]
-#[Polyglot\Locale(primary: 'en_GB')]
-class TranslatingUnmappedPropertiesTest_Entity
-{
-    #[ORM\Column(type: 'integer')]
-    #[ORM\Id]
-    #[ORM\GeneratedValue]
-    public ?int $id = null;
-
-    #[ORM\OneToMany(targetEntity: TranslatingUnmappedPropertiesTest_Translation::class, mappedBy: 'entity')]
-    #[Polyglot\TranslationCollection]
-    protected Collection $translations;
-
-    #[ORM\Column(type: 'string')]
-    public $mappedText;
-
-    // (!) This field is unmapped from the ORM point of view
-    #[Polyglot\Translatable]
-    public $text;
-
-    public function __construct()
-    {
-        $this->translations = new ArrayCollection();
-        $this->text = new Translatable();
-    }
-
-    #[ORM\PreFlush]
-    public function copyToMappedField(): void
-    {
-        $this->mappedText = $this->text;
-    }
-
-    #[ORM\PostLoad]
-    public function copyFromMappedField(): void
-    {
-        $this->text = $this->mappedText;
-    }
-}
-
-#[ORM\Entity]
-#[ORM\HasLifecycleCallbacks]
-class TranslatingUnmappedPropertiesTest_Translation
-{
-    #[ORM\Column(type: 'integer')]
-    #[ORM\Id]
-    #[ORM\GeneratedValue]
-    private ?int $id = null;
-
-    #[ORM\Column(type: 'string')]
-    #[Polyglot\Locale]
-    private string $locale;
-
-    #[ORM\ManyToOne(targetEntity: TranslatingUnmappedPropertiesTest_Entity::class, inversedBy: 'translations')]
-    private TranslatingUnmappedPropertiesTest_Entity $entity;
-
-    #[ORM\Column(type: 'string')]
-    private $mappedText;
-
-    // (!) This field is unmapped from the ORM point of view
-    private $text;
-
-    #[ORM\PreFlush]
-    public function copyToMappedField(): void
-    {
-        $this->mappedText = $this->text;
-    }
-
-    #[ORM\PostLoad]
-    public function copyFromMappedField(): void
-    {
-        $this->text = $this->mappedText;
     }
 }
