+++ tests/Functional/TranslationPropertyNamedDifferentlyTest.php
@@ -2,23 +2,21 @@
 
 namespace Webfactory\Bundle\PolyglotBundle\Tests\Functional;
 
-use Doctrine\Common\Collections\ArrayCollection;
-use Doctrine\Common\Collections\Collection;
-use Doctrine\ORM\Mapping as ORM;
-use Webfactory\Bundle\PolyglotBundle\Attribute as Polyglot;
+use Webfactory\Bundle\PolyglotBundle\Tests\Fixtures\Entity\TranslationPropertyNamedDifferently\TranslationPropertyNamedDifferently_Entity;
+use Webfactory\Bundle\PolyglotBundle\Tests\Fixtures\Entity\TranslationPropertyNamedDifferently\TranslationPropertyNamedDifferently_Translation;
 use Webfactory\Bundle\PolyglotBundle\Translatable;
-use Webfactory\Bundle\PolyglotBundle\TranslatableInterface;
 
 /**
  * This tests a setup where the "translation" field is named different from the
  * field in the base entity class.
  */
-class TranslationPropertyNamedDifferentlyTest extends FunctionalTestBase
+class TranslationPropertyNamedDifferentlyTest extends DatabaseFunctionalTestCase
 {
     protected function setUp(): void
     {
         parent::setUp();
-        $this->setupOrmInfrastructure([
+
+        self::setupSchema([
             TranslationPropertyNamedDifferently_Entity::class,
             TranslationPropertyNamedDifferently_Translation::class,
         ]);
@@ -31,9 +29,9 @@ class TranslationPropertyNamedDifferentlyTest extends FunctionalTestBase
         $translatable->setTranslation('Basistext', 'de_DE');
         $entity->setText($translatable);
 
-        $this->infrastructure->import($entity);
+        self::import($entity);
 
-        $loaded = $this->infrastructure->getEntityManager()->find(TranslationPropertyNamedDifferently_Entity::class, $entity->getId());
+        $loaded = $this->entityManager->find(TranslationPropertyNamedDifferently_Entity::class, $entity->getId());
 
         self::assertSame('Basistext', $loaded->getText()->translate('de_DE'));
         self::assertSame('base text', $loaded->getText()->translate('en_GB'));
@@ -41,10 +39,10 @@ class TranslationPropertyNamedDifferentlyTest extends FunctionalTestBase
 
     public function testAddTranslation(): void
     {
-        $entityManager = $this->infrastructure->getEntityManager();
+        $entityManager = $this->entityManager;
         $entity = new TranslationPropertyNamedDifferently_Entity();
         $entity->setText(new Translatable('base text'));
-        $this->infrastructure->import($entity);
+        self::import($entity);
 
         $loaded = $entityManager->find(TranslationPropertyNamedDifferently_Entity::class, $entity->getId());
         $loaded->getText()->setTranslation('Basistext', 'de_DE');
@@ -59,13 +57,13 @@ class TranslationPropertyNamedDifferentlyTest extends FunctionalTestBase
 
     public function testUpdateTranslations(): void
     {
-        $entityManager = $this->infrastructure->getEntityManager();
+        $entityManager = $this->entityManager;
 
         $entity = new TranslationPropertyNamedDifferently_Entity();
         $translatable = new Translatable('base text');
         $translatable->setTranslation('Basistext', 'de_DE');
         $entity->setText($translatable);
-        $this->infrastructure->import($entity);
+        self::import($entity);
 
         $loaded = $entityManager->find(TranslationPropertyNamedDifferently_Entity::class, $entity->getId());
         $loaded->getText()->setTranslation('new base text');
@@ -78,61 +76,4 @@ class TranslationPropertyNamedDifferentlyTest extends FunctionalTestBase
         self::assertSame('new base text', $reloaded->getText()->translate('en_GB'));
         self::assertSame('neuer Basistext', $reloaded->getText()->translate('de_DE'));
     }
-}
-
-#[Polyglot\Locale(primary: 'en_GB')]
-#[ORM\Entity]
-class TranslationPropertyNamedDifferently_Entity
-{
-    #[ORM\Column(type: 'integer')]
-    #[ORM\Id]
-    #[ORM\GeneratedValue]
-    private ?int $id = null;
-
-    #[Polyglot\TranslationCollection]
-    #[ORM\OneToMany(targetEntity: \TranslationPropertyNamedDifferently_Translation::class, mappedBy: 'entity')]
-    protected Collection $translations;
-
-    #[Polyglot\Translatable(translationFieldname: 'textOtherName')]
-    #[ORM\Column(type: 'string')]
-    protected string|TranslatableInterface|null $text = null;
-
-    public function __construct()
-    {
-        $this->translations = new ArrayCollection();
-    }
-
-    public function getId(): int
-    {
-        return $this->id;
-    }
-
-    public function setText(TranslatableInterface $text): void
-    {
-        $this->text = $text;
-    }
-
-    public function getText(): ?TranslatableInterface
-    {
-        return $this->text;
-    }
-}
-
-#[ORM\Entity]
-class TranslationPropertyNamedDifferently_Translation
-{
-    #[ORM\Id]
-    #[ORM\GeneratedValue]
-    #[ORM\Column(type: 'integer')]
-    private ?int $id = null;
-
-    #[Polyglot\Locale]
-    #[ORM\Column]
-    private string $locale;
-
-    #[ORM\ManyToOne(targetEntity: \TranslationPropertyNamedDifferently_Entity::class, inversedBy: 'translations')]
-    private TranslationPropertyNamedDifferently_Entity $entity;
-
-    #[ORM\Column]
-    private string $textOtherName;
 }
